tipps changes, merge & refactoring
