+++ server/gokbg3/grails-app/controllers/org/gokb/IntegrationController.groovy
@@ -22,6 +22,7 @@ class IntegrationController {
   def concurrencyManagerService
   def classExaminationService
   def componentUpdateService
+  def componentLookupService
   def titleLookupService
   def applicationEventService
   def reviewRequestService
@@ -228,7 +229,7 @@ class IntegrationController {
     def id = request.JSON.'@id'
     def new_org = new Org(name:name)
 
-    def primary_identifier = Identifier.lookupOrCreateCanonicalIdentifier('global',id)
+    def primary_identifier = componentLookupService.lookupOrCreateCanonicalIdentifier('global',id)
     new_org.ids.add(primary_identifier)
 
     request.JSON.'owl:sameAs'?.each {  said ->
@@ -236,7 +237,7 @@ class IntegrationController {
       // Double check that this identifier is NOT already used
       def existing_usage = KBComponent.lookupByIO('global',said)
       if ( existing_usage == null ) {
-        def identifier = Identifier.lookupOrCreateCanonicalIdentifier('global',said)
+        def identifier = componentLookupService.lookupOrCreateCanonicalIdentifier('global',said)
         new_org.ids.add(identifier)
       }
       else {
@@ -679,7 +680,7 @@ class IntegrationController {
       if ( ci.type && ci.value && ci.type.toLowerCase() != "originediturl") {
 
         if (!ids.contains(testKey)) {
-          def canonical_identifier = Identifier.lookupOrCreateCanonicalIdentifier(ci.type,ci.value)
+          def canonical_identifier = componentLookupService.lookupOrCreateCanonicalIdentifier(ci.type,ci.value)
 
           log.debug("Checking identifiers of component ${component.id}")
 
@@ -935,11 +936,11 @@ class IntegrationController {
 
           job.ownerId = user.id
 
-          def pkg_validation = Package.validateDTO(json.packageHeader)
+          def pkg_validation = packageService.validateDTO(json.packageHeader)
 
           if ( pkg_validation.valid ) {
             try {
-              def the_pkg = Package.upsertDTO(json.packageHeader, user)
+              def the_pkg = packageService.upsertDTO(json.packageHeader, user)
               def existing_tipps = []
               def valid = true
               Boolean curated_pkg = false;
@@ -976,9 +977,10 @@ class IntegrationController {
 
                   Map platform_cache = [:]
                   log.debug("\n\n\nPackage ID: ${the_pkg.id} / ${json.packageHeader}");
+                  def idx = 0
 
                   // Validate and upsert titles and platforms
-                  json.tipps.eachWithIndex { json_tipp, idx ->
+                  for (json_tipp in json.tipps) {
 
                     def title_validation = TitleInstance.validateDTO(json_tipp.title);
                     def tipp_plt_dto = json_tipp.hostPlatform ?: json_tipp.platform
@@ -1067,7 +1069,7 @@ class IntegrationController {
 
                             titleLookupService.addPublisherHistory(ti, titleObj.publisher_history)
 
-                            ti.save(flush:true)
+                            ti = ti.merge(flush:true)
 
                             json_tipp.title.internalId = ti.id
                           } else {
@@ -1181,6 +1183,15 @@ class IntegrationController {
                     if (idx % 50 == 0) {
                       cleanUpGorm()
                     }
+
+                    if ( Thread.currentThread().isInterrupted() ) {
+                      log.debug("Job cancelling ..")
+                      job_result.result = "CANCELLED"
+                      job_result.messageService = "Import was cancelled before completion!"
+                      break;
+                    }
+
+                    idx++
                     job.setProgress(idx, json.tipps.size() * 2)
                   }
                 }
@@ -1195,9 +1206,10 @@ class IntegrationController {
 
                 int tippctr=0;
                 if ( valid ) {
+                  def vidx = 0
                   // If valid so far, validate tipps
                   log.debug("Validating tipps [${tippctr++}]");
-                  json.tipps.eachWithIndex { json_tipp, idx ->
+                  for (json_tipp in json.tipps) {
                     def validation_result = TitleInstancePackagePlatform.validateDTO(json_tipp)
 
                     if ( validation_result && !validation_result.valid ) {
@@ -1205,7 +1217,7 @@ class IntegrationController {
                       valid = false
                       def tipp_errors = [
                         'code': 400,
-                        idx: idx,
+                        idx: vidx,
                         message: messageService.resolveCode('crossRef.package.tipps.error.preValidation', [json_tipp.title.name, validation_result.errors], locale),
                         baddata: json_tipp,
                         errors: validation_result.errors
@@ -1213,9 +1225,18 @@ class IntegrationController {
                       errors.tipps.add(tipp_errors)
                     }
 
-                    if (idx % 50 == 0) {
+                    if (vidx % 50 == 0) {
                       cleanUpGorm()
                     }
+
+                    if ( Thread.currentThread().isInterrupted() ) {
+                      log.debug("Job cancelling ..")
+                      job_result.result = "CANCELLED"
+                      job_result.message = "Import was cancelled before completion!"
+                      break;
+                    }
+
+                    vidx++
                   }
                 }
                 else {
@@ -1235,6 +1256,7 @@ class IntegrationController {
 
                   def tipp_upsert_start_time = System.currentTimeMillis()
                   def tipp_fails = 0
+                  def tidx = 0
 
                   if ( json.tipps?.size() > 0 ) {
                     Package.withNewSession {
@@ -1249,18 +1271,18 @@ class IntegrationController {
                   }
 
                   // If valid, upsert tipps
-                  json.tipps.eachWithIndex { tipp, idx ->
+                  for(json_tipp in json.tipps) {
                     tippctr++
 
-                    log.debug("Upsert tipp [${tippctr}] ${tipp}")
+                    log.debug("Upsert tipp [${tippctr}] ${json_tipp}")
                     def upserted_tipp = null
 
                     try {
-                      upserted_tipp = TitleInstancePackagePlatform.upsertDTO(tipp, user)
+                      upserted_tipp = TitleInstancePackagePlatform.upsertDTO(json_tipp, user)
                       log.debug("Upserted TIPP ${upserted_tipp} with URL ${upserted_tipp?.url}")
                       upserted_tipp = upserted_tipp?.merge(flush: true)
 
-                      componentUpdateService.ensureCoreData(upserted_tipp, tipp, fullsync)
+                      componentUpdateService.ensureCoreData(upserted_tipp, json_tipp, fullsync)
                     }
                     catch (grails.validation.ValidationException ve) {
                       log.error("ValidationException attempting to cross reference TIPP",ve);
@@ -1268,9 +1290,9 @@ class IntegrationController {
                       tipp_fails++
                       def tipp_errors = [
                         code: 400,
-                        idx: idx,
-                        message: messageService.resolveCode('crossRef.package.tipps.error.validation', [tipp.title.name], locale),
-                        baddata: tipp,
+                        idx: tidx,
+                        message: messageService.resolveCode('crossRef.package.tipps.error.validation', [json_tipp.title.name], locale),
+                        baddata: json_tipp,
                         errors: messageService.processValidationErrors(ve.errors)
                       ]
                       errors.tipps.add(tipp_errors)
@@ -1284,9 +1306,9 @@ class IntegrationController {
                       tipp_fails++
                       def tipp_errors = [
                         code: 500,
-                        idx: idx,
-                        message: messageService.resolveCode('crossRef.package.tipps.error', [tipp.title.name], locale),
-                        baddata: tipp
+                        idx: tidx,
+                        message: messageService.resolveCode('crossRef.package.tipps.error', [json_tipp.title.name], locale),
+                        baddata: json_tipp
                       ]
                       errors.tipps.add(tipp_errors)
 
@@ -1297,21 +1319,21 @@ class IntegrationController {
                     if (upserted_tipp) {
                       if ( existing_tipps.size() > 0 && upserted_tipp && existing_tipps.contains(upserted_tipp.id) ) {
                         log.debug("Existing TIPP matched!")
-                        tipps_to_delete.remove(upserted_tipp.id)
+                        tipps_to_delete.removeElement(upserted_tipp.id)
                       }
 
-                      if ( upserted_tipp && upserted_tipp?.status != status_deleted && tipp.status == "Deleted" ) {
+                      if ( upserted_tipp && upserted_tipp?.status != status_deleted && json_tipp.status == "Deleted" ) {
                         upserted_tipp.deleteSoft()
                         num_removed_tipps++;
                       }
-                      else if ( upserted_tipp && upserted_tipp?.status != status_retired && tipp.status == "Retired" ) {
+                      else if ( upserted_tipp && upserted_tipp?.status != status_retired && json_tipp.status == "Retired" ) {
                         upserted_tipp.retire()
                         num_removed_tipps++;
                       }
-                      else if ( upserted_tipp && upserted_tipp.status != status_current && (!tipp.status || tipp.status == "Current") ) {
+                      else if ( upserted_tipp && upserted_tipp.status != status_current && (!json_tipp.status || json_tipp.status == "Current") ) {
                         upserted_tipp.setActive()
 
-                        if (upserted_tipp.isDeleted() && status_current) {
+                        if (upserted_tipp.isDeleted()) {
                             reviewRequestService.raise(
                             upserted_tipp,
                             "Matched TIPP was marked as Deleted.",
@@ -1346,17 +1368,26 @@ class IntegrationController {
                       tipp_fails++
                       def tipp_errors = [
                         code: 500,
-                        idx: idx,
-                        message: messageService.resolveCode('crossRef.package.tipps.error', [tipp.title.name], locale),
-                        baddata: tipp
+                        idx: tidx,
+                        message: messageService.resolveCode('crossRef.package.tipps.error', [json_tipp.title.name], locale),
+                        baddata: json_tipp
                       ]
                       errors.tipps.add(tipp_errors)
                     }
 
-                    if (idx % 50 == 0) {
+                    if (tidx % 50 == 0) {
                       cleanUpGorm()
                     }
-                    job.setProgress(idx + json.tipps.size(), json.tipps.size() * 2)
+
+                    if ( Thread.currentThread().isInterrupted() ) {
+                      log.debug("Job cancelling ..")
+                      job_result.result = "CANCELLED"
+                      job_result.message = "Import was cancelled before completion!"
+                      break;
+                    }
+
+                    tidx++
+                    job.setProgress(tidx + json.tipps.size(), json.tipps.size() * 2)
                   }
 
                   if (!valid) {
@@ -2069,7 +2100,7 @@ class IntegrationController {
     book_changed
   }
 
-  @Secured(['ROLE_API', 'IS_AUTHENTICATED_FULLY'])
+  @Secured(value=["hasRole('ROLE_API')", 'IS_AUTHENTICATED_FULLY'])
   def getJobInfo() {
     def result = [ 'result' : 'OK', 'params' : params ]
     User user = springSecurityService.currentUser
