+++ server/gokbg3/grails-app/domain/org/gokb/cred/Org.groovy
@@ -342,116 +342,4 @@ class Org extends KBComponent {
     Combo.executeUpdate("delete from Combo where fromComponent.id = ?", [this.getId()]);
     result
   }
-
-  @Transient
-  public static Org upsertDTO(orgDTO, def user = null) {
-    log.info("Upsert package with header ${orgDTO}");
-    def status_deleted = RefdataCategory.lookupOrCreate('KBComponent.Status', 'Deleted')
-    def org_normname = Org.generateNormname(orgDTO.name)
-
-    log.debug("Checking by normname ${org_normname} ..")
-    def name_candidates = Org.executeQuery("from Org as p where p.normname = ? and p.status <> ?", [org_normname, status_deleted])
-    def full_matches = []
-    def created = false
-    def result = orgDTO.uuid ? Org.findByUuid(orgDTO.uuid) : null;
-    boolean changed = false;
-
-    if (!result && name_candidates.size() > 0 && orgDTO.identifiers?.size() > 0) {
-      log.debug("Got ${name_candidates.size()} matches by name. Checking against identifiers!")
-      name_candidates.each { mp ->
-        if (mp.ids.size() > 0) {
-          def id_match = false;
-
-          orgDTO.identifiers.each { rid ->
-
-            Identifier the_id = Identifier.lookupOrCreateCanonicalIdentifier(rid.type, rid.value);
-
-            if (mp.ids.contains(the_id)) {
-              id_match = true
-            }
-          }
-
-          if (id_match && !full_matches.contains(mp)) {
-            full_matches.add(mp)
-          }
-        }
-      }
-
-      if (full_matches.size() == 1) {
-        log.debug("Matched org by name + identifier!")
-        result = full_matches[0]
-      } else if (full_matches.size() == 0 && name_candidates.size() == 1) {
-        result = name_candidates[0]
-        log.debug("Found a single match by name!")
-      } else {
-        log.warn("Found multiple possible matches for org! Aborting..")
-        return result
-      }
-    } else if (!result && name_candidates.size() == 1) {
-      log.debug("Matched org by name!")
-      result = name_candidates[0]
-    } else if (result && result.name != orgDTO.name) {
-      def current_name = result.name
-      changed |= ClassUtils.setStringIfDifferent(result, 'name', orgDTO.name)
-
-      if (!result.variantNames.find { it.variantName == current_name }) {
-        def new_variant = new KBComponentVariantName(owner: result, variantName: current_name).save(flush: true, failOnError: true)
-      }
-    }
-
-    if (!result) {
-      log.debug("Did not find a match via name, trying existing variantNames..")
-      def variant_normname = GOKbTextUtils.normaliseString(orgDTO.name)
-      def variant_candidates = Org.executeQuery("select distinct p from Org as p join p.variantNames as v where v.normVariantName = ? and p.status <> ? ", [variant_normname, status_deleted]);
-
-      if (variant_candidates.size() == 1) {
-        result = variant_candidates[0]
-        log.debug("Package matched via existing variantName.")
-      }
-    }
-
-    if (!result && orgDTO.variantNames?.size() > 0) {
-      log.debug("Did not find a match via existing variantNames, trying supplied variantNames..")
-      orgDTO.variantNames.each {
-
-        if (it.trim().size() > 0) {
-          result = Org.findByName(it)
-
-          if (result) {
-            log.debug("Found existing package name for variantName ${it}")
-          } else {
-
-            def variant_normname = GOKbTextUtils.normaliseString(it)
-            def variant_candidates = Org.executeQuery("select distinct p from Org as p join p.variantNames as v where v.normVariantName = ? and p.status <> ? ", [variant_normname, status_deleted]);
-
-            if (variant_candidates.size() == 1) {
-              log.debug("Found existing Org variant name for variantName ${it}")
-              result = variant_candidates[0]
-            }
-          }
-        }
-      }
-    }
-
-    if (!result) {
-      log.debug("No existing Org matched. Creating new Org..")
-
-      result = new Org(name: orgDTO.name, normname: org_normname)
-
-      created = true
-
-      if (orgDTO.uuid && orgDTO.uuid.trim().size() > 0) {
-        result.uuid = orgDTO.uuid
-      }
-
-      result.save(flush: true, failOnError: true)
-    } else if (user && !user.hasRole('ROLE_SUPERUSER') && result.curatoryGroups && result.curatoryGroups?.size() > 0) {
-      def cur = user.curatoryGroups?.id.intersect(result.curatoryGroups?.id)
-
-      if (!cur) {
-        log.debug("No curator!")
-        return result
-      }
-    }
-  }
 }
