+++ server/gokbg3/src/integration-test/groovy/gokbg3/rest/SourcesTestSpec.groovy
@@ -38,7 +38,7 @@ class SourcesTestSpec extends AbstractAuthSpec {
     }
     then:
     resp.status == 200
-    resp.json.data.size() == 8
+    resp.json.data.size() == 9
   }
 
   void "test GET /rest/sources/{id}"() {
